requirements.txt
aniso8601==8.0.0
api==0.0.7
appdirs==1.4.4
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-texttospeech==2.2.0
google-cloud-translate==3.0.1
googleapis-common-protos==1.52.0
grpcio==1.33.2
identify==1.5.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
JsonForm==0.0.2
jsonschema==3.2.0
JsonSir==0.0.2
libcst==0.3.13
lxml==4.6.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nodeenv==1.5.0
nose==1.3.7
packaging==20.4
pluggy==0.13.1
pre-commit==2.8.2
proto-plus==1.11.0
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.17.3
Python-EasyConfig==0.1.7
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
Resource==0.2.1
rsa==4.6
six==1.15.0
toml==0.10.2
tox==3.20.1
translate==3.5.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
virtualenv==20.1.0
Werkzeug==1.0.1